refactor: move OsPlacesDpaDto to OsPlaces folder
